+++ patches/src/main/kotlin/app/revanced/patches/all/misc/shortcut/sharetargets/RemoveShareTargetsPatch.kt
@@ -17,7 +17,8 @@ val removeShareTargetsPatch = resourcePatch(
         try {
             document("res/xml/shortcuts.xml")
         } catch (_: FileNotFoundException) {
-            return@execute Logger.getLogger(this::class.java.name).warning("The app has no shortcuts")
+            return@execute Logger.getLogger(this::class.java.name).warning(
+                "The app has no shortcuts.  No changes applied.")
         }.use { document ->
             val rootNode = document.getNode("shortcuts") as? Element ?: return@use