+++ patches/src/main/kotlin/app/revanced/patches/yuka/misc/unlockpremium/UnlockPremiumPatch.kt
@@ -3,10 +3,10 @@ package app.revanced.patches.yuka.misc.unlockpremium
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
 import app.revanced.patcher.patch.bytecodePatch
 
+@Deprecated("This patch no longer works and will be removed in the future.")
 @Suppress("unused")
-val unlockPremiumPatch = bytecodePatch(
-    name = "Unlock premium",
-) {
+val unlockPremiumPatch = bytecodePatch {
+
     compatibleWith("io.yuka.android"("4.29"))
 
     execute {